+++ README.md
@@ -1,4 +1,4 @@
-# Stepit
+# Steppie
 
 [![Deploy](https://www.herokucdn.com/deploy/button.png)](https://heroku.com/deploy)
