+++ app/config/parameters.yml.dist
@@ -19,4 +19,4 @@ parameters:
     # A secret key that's used to generate certain security-related tokens
     secret:            ThisTokenIsNotSoSecretChangeIt
 
-    site_name:         Stepit
+    site_name:         Steppie
