+++ app/config/routing.yml
@@ -1,5 +1,5 @@
-stepit_app:
-    resource: "@StepitAppBundle/Resources/config/routing/routing.yml"
+steppie_app:
+    resource: "@SteppieAppBundle/Resources/config/routing/routing.yml"
     prefix:   /
 
 easy_admin:
