--- src/Stepit/Bundle/AppBundle/Resources/config/routing/rest.yml
@@ -1,7 +1,0 @@
-stepit_app_rest_content:
-  resource: "@StepitAppBundle/Controller/Rest/ContentController.php"
-  type: rest
-  prefix: api/v1
-  name_prefix: api_v1_stepit_app_
-  options:
-    expose: true
