--- src/Stepit/Bundle/AppBundle/Resources/public/js/script.js
@@ -1,26 +1,0 @@
-$(function () {
-    $('table.table-stepit').stickyTableHeaders();
-
-    $('.btn-done').on('click', function () {
-        var matterId = $(this).data('matter-id');
-        var stepId = $(this).data('step-id');
-        var now = new Date;
-        var todayString = now.getFullYear() + '/' + (now.getMonth() + 1) + '/' + now.getDate();
-
-        $.ajax({
-            type: 'POST',
-            async: true,
-            url: Routing.generate('api_v1_stepit_app_post_content', { _format: 'json' }),
-            data: {
-                value: todayString,
-                matter: matterId,
-                step: stepId
-            },
-            dataType: 'json',
-            context: this,
-            success: function (data) {
-                $(this).closest('td').text(data.value);
-            }
-        });
-    });
-});
