--- src/Stepit/Bundle/AppBundle/Resources/views/Default/project.html.twig
@@ -1,89 +1,0 @@
-{% extends 'StepitAppBundle::layout.html.twig' %}
-
-{% block page_header %}
-    <ol class="breadcrumb">
-        <li><a href="{{ path('home') }}">{{ 'project.list' | trans }}</a></li>
-        <li class="active">{{ project.name }}</li>
-    </ol>
-{% endblock %}
-
-{% block content_row %}
-    <div class="table-responsive">
-        <table class="table table-hover table-stepit">
-            <thead>
-            <tr>
-                <th>{{ 'matter.id' | trans }}</th>
-                {% for matter in matters %}
-                    <td>{{ matter.id }}</td>
-                {% endfor %}
-            </tr>
-            <tr>
-                <th>{{ 'matter.name' | trans }}</th>
-                {% for matter in matters %}
-                    <td>
-                        {{ matter.name }}
-                        {% if matter.description is not empty %}
-                            <span data-toggle="tooltip" data-placement="top" title="{{ matter.description }}">
-                                <i class="fa fa-info-circle"></i>
-                            </span>
-                        {% endif %}
-                    </td>
-                {% endfor %}
-            </tr>
-            <tr>
-                <th>{{ 'matter.owners' | trans }}</th>
-                {% for matter in matters %}
-                    <td>
-                        {% for owner in matter.owners %}
-                            <span class="label label-default label-xs">{{ owner }}</span>
-                        {% endfor %}
-                    </td>
-                {% endfor %}
-            </tr>
-            </thead>
-
-            <tbody>
-            {% for step in steps %}
-                <tr>
-                    <th>
-                        {% if step.link is not empty %}
-                            <a href="{{ step.link }}" target="_blank">{{ step.name }}</a>
-                        {% else %}
-                            {{ step.name }}
-                        {% endif %}
-                        {% if step.description is not empty %}
-                            <span data-toggle="tooltip" data-placement="top" title="{{ step.description }}">
-                                <i class="fa fa-info-circle"></i>
-                            </span>
-                        {% endif %}
-                    </th>
-                    {% for matter in matters %}
-                        <td>
-                            {% if contents[matter.id][step.id] is defined %}
-                                {{ contents[matter.id][step.id].value }}
-                            {% else %}
-                                <p class="btn btn-success btn-xs btn-done" data-matter-id="{{ matter.id }}" data-step-id="{{ step.id }}">{{ 'done' | trans }}</p>
-                            {% endif %}
-                        </td>
-                    {% endfor %}
-                </tr>
-            {% endfor %}
-            </tbody>
-        </table>
-    </div>
-{% endblock %}
-
-{% block foot_script_assetic %}
-    {% javascripts
-    '//ajax.googleapis.com/ajax/libs/jquery/1.11.1/jquery.min.js'
-    '@MopaBootstrapBundle/Resources/public/bootstrap/dist/js/bootstrap.js'
-    '@MopaBootstrapBundle/Resources/public/js/mopabootstrap-collection.js'
-    '@MopaBootstrapBundle/Resources/public/js/mopabootstrap-subnav.js'
-    'https://raw.githubusercontent.com/jmosbech/StickyTableHeaders/master/js/jquery.stickytableheaders.min.js'
-    '@StepitAppBundle/Resources/public/js/script.js'
-    %}
-    <script type="text/javascript" src="{{ asset_url }}"></script>
-    <script src="{{ asset('bundles/fosjsrouting/js/router.js') }}"></script>
-    <script src="{{ path('fos_js_routing_js', { 'callback': 'fos.Router.setData' }) }}"></script>
-    {% endjavascripts %}
-{% endblock %}
