--- src/Stepit/Bundle/AppBundle/StepitAppBundle.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Stepit\Bundle\AppBundle;
-
-use Symfony\Component\HttpKernel\Bundle\Bundle;
-
-class StepitAppBundle extends Bundle
-{
-}
