+++ src/Steppie/Bundle/AppBundle/Entity/Repository/MatterRepository.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Steppie\Bundle\AppBundle\Entity\Repository;
+
+use Doctrine\ORM\EntityRepository;
+
+/**
+ * MatterRepository
+ *
+ * This class was generated by the Doctrine ORM. Add your own custom
+ * repository methods below.
+ */
+class MatterRepository extends EntityRepository
+{
+}
