+++ src/Steppie/Bundle/AppBundle/Entity/Repository/StepRepository.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Steppie\Bundle\AppBundle\Entity\Repository;
+
+use Doctrine\ORM\EntityRepository;
+
+/**
+ * StepRepository
+ *
+ * This class was generated by the Doctrine ORM. Add your own custom
+ * repository methods below.
+ */
+class StepRepository extends EntityRepository
+{
+}
