+++ src/Steppie/Bundle/AppBundle/Resources/config/routing/rest.yml
@@ -1,0 +1,7 @@
+steppie_app_rest_content:
+  resource: "@SteppieAppBundle/Controller/Rest/ContentController.php"
+  type: rest
+  prefix: api/v1
+  name_prefix: api_v1_steppie_app_
+  options:
+    expose: true
