+++ src/Steppie/Bundle/AppBundle/Resources/public/less/style.less
@@ -1,0 +1,25 @@
+@import "../../../../../../../vendor/mopa/bootstrap-bundle/Mopa/Bundle/BootstrapBundle/Resources/public/less/mopabootstrapbundle-fa4";
+@fa-font-path:  "/bundles/mopabootstrap/fonts/fa4";
+
+table.table-steppie {
+    thead {
+        border-bottom: 3px double #ddd;
+        background-color: #fff;
+        th {
+            border-bottom: none;
+            border-right: 3px double #ddd;
+        }
+    }
+    tbody {
+        th {
+            border-right: 3px double #ddd;
+        }
+    }
+}
+
+footer {
+    border-top: 1px solid #ddd;
+    padding: 20px 0;
+    margin-top: 50px;
+    text-align: center;
+}
