+++ src/Steppie/Bundle/AppBundle/Resources/views/layout.html.twig
@@ -1,0 +1,48 @@
+{% extends 'MopaBootstrapBundle::base.html.twig' %}
+
+{% block title %}
+    {% block subtitle %}{% endblock %}
+    {% if block('subtitle') is not empty %} | {% endif %}
+    {{ site_name }}
+{% endblock %}
+
+{% block navbar %}
+    {% embed '@MopaBootstrap/Navbar/navbar.html.twig' with { fixedTop: false, staticTop: true, inverse: false, fluid: true } %}
+        {% block brand %}
+            <a class="navbar-brand" href="{{ path('home') }}">{{ site_name }}</a>
+        {% endblock %}
+        {% block menu %}
+            <ul class="nav navbar-nav navbar-right">
+                <li><a href="{{ path('admin') }}" target="_blank">{{ 'admin' | trans }}</a></li>
+            </ul>
+        {% endblock %}
+    {% endembed %}
+{% endblock navbar %}
+
+{% block container_class %}container-fluid{% endblock container_class %}
+
+{% block page_header %}{% endblock page_header %}
+
+{% block footer %}
+    <p>Powered by <a href="https://github.com/qckanemoto/Steppie" target="_blank">Steppie</a></p>
+{% endblock footer %}
+
+{% block head_style %}
+    {% stylesheets filter='less,cssrewrite'
+    '@SteppieAppBundle/Resources/public/less/style.less'
+    %}
+    <link href="{{ asset_url }}" type="text/css" rel="stylesheet" media="screen" />
+    {% endstylesheets %}
+{% endblock %}
+
+{% block foot_script_assetic %}
+    {% javascripts
+    '//ajax.googleapis.com/ajax/libs/jquery/1.11.1/jquery.min.js'
+    '@MopaBootstrapBundle/Resources/public/bootstrap/dist/js/bootstrap.js'
+    '@MopaBootstrapBundle/Resources/public/js/mopabootstrap-collection.js'
+    '@MopaBootstrapBundle/Resources/public/js/mopabootstrap-subnav.js'
+    '@SteppieAppBundle/Resources/public/js/script.js'
+    %}
+    <script type="text/javascript" src="{{ asset_url }}"></script>
+    {% endjavascripts %}
+{% endblock %}
