+++ src/Steppie/Bundle/AppBundle/SteppieAppBundle.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Steppie\Bundle\AppBundle;
+
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+class SteppieAppBundle extends Bundle
+{
+}
